Add MongoDriverInformation when creating the MongoClient
